tasks: move ui bits to separate module
